PWN-783: Move canvas elements
