+++ lib/multi_version_common_cartridge.rb
@@ -24,8 +24,8 @@ require 'multi_version_common_cartridge/resources/resource'
 require 'multi_version_common_cartridge/resources/basic_lti_link'
 require 'multi_version_common_cartridge/resources/canvas_assignment'
 require 'multi_version_common_cartridge/resources/canvas_course_settings'
-require 'multi_version_common_cartridge/resources/canvas_elements/assignment'
-require 'multi_version_common_cartridge/resources/canvas_elements/course_settings'
+require 'multi_version_common_cartridge/elements/canvas/assignment'
+require 'multi_version_common_cartridge/elements/canvas/course_settings'
 require 'multi_version_common_cartridge/manifest'
 require 'multi_version_common_cartridge/cartridge'
 require 'multi_version_common_cartridge/xml_definitions'
